+++ glean-core/android/src/main/java/mozilla/telemetry/glean/Glean.kt
@@ -2,6 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+@file:Suppress("ktlint:standard:no-wildcard-imports")
+
 package mozilla.telemetry.glean
 
 import android.app.ActivityManager
@@ -16,11 +18,12 @@ import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.launch
 import mozilla.telemetry.glean.GleanMetrics.GleanValidation
 import mozilla.telemetry.glean.config.Configuration
-import mozilla.telemetry.glean.internal.* // ktlint-disable no-wildcard-imports
+import mozilla.telemetry.glean.internal.*
 import mozilla.telemetry.glean.net.BaseUploader
 import mozilla.telemetry.glean.scheduler.GleanLifecycleObserver
 import mozilla.telemetry.glean.scheduler.MetricsPingScheduler
 import mozilla.telemetry.glean.scheduler.PingUploadWorker
+import mozilla.telemetry.glean.scheduler.PingUploadWorker.Companion.performUpload
 import mozilla.telemetry.glean.utils.ThreadUtils
 import mozilla.telemetry.glean.utils.calendarToDatetime
 import mozilla.telemetry.glean.utils.canWriteToDatabasePath
@@ -57,7 +60,16 @@ internal class OnGleanEventsImpl(val glean: GleanInternalAPI) : OnGleanEvents {
     }
 
     override fun triggerUpload() {
-        PingUploadWorker.enqueueWorker(glean.applicationContext)
+        if (!glean.isCustomDataPath) {
+            PingUploadWorker.enqueueWorker(glean.applicationContext)
+        } else {
+            // WorkManager wants to run on the main thread/process typically, so when Glean is
+            // running in a background process we will instead just use the internal Glean
+            // coroutine dispatcher to run the upload task.
+            Dispatchers.API.executeTask {
+                performUpload()
+            }
+        }
     }
 
     override fun startMetricsPingScheduler(): Boolean {