+++ samples/android/app/src/main/java/org/mozilla/samples/gleancore/MainActivity.kt
@@ -4,6 +4,7 @@
 
 package org.mozilla.samples.gleancore
 
+import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
@@ -15,10 +16,16 @@ import org.mozilla.samples.gleancore.databinding.ActivityMainBinding
 
 open class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
+    private lateinit var serviceIntent: Intent
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
+        serviceIntent = Intent(this, SampleBackgroundProcess::class.java)
+
+        // Start the background service to test recording metrics and sending pings in from off
+        // the main process
+        startService(serviceIntent)
 
         setContentView(binding.root)
 
@@ -72,5 +79,10 @@ open class MainActivity : AppCompatActivity() {
         }
 
         Test.timespan.stop()
+    }
+
+    override fun onDestroy() {
+        stopService(serviceIntent)
+        super.onDestroy()
     }
 }