+++ samples/rust/src/main.rs
@@ -45,6 +45,7 @@ fn main() {
         app_build: env!("CARGO_PKG_VERSION").to_string(),
         app_display_version: env!("CARGO_PKG_VERSION").to_string(),
         channel: None,
+        locale: None,
     };
 
     glean::initialize(cfg, client_info);