+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v53.1.0...main)
+[Full changelog](https://github.com/mozilla/glean/compare/v53.2.0...main)
+
+# v53.2.0 (2023-08-02)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v53.1.0...v53.2.0)
+
+* General
+  * Update `glean_parser` to v8.1.0. Subsequently, metric names now have a larger limit of 70 characters ([release notes](https://github.com/mozilla/glean_parser/releases/tag/v8.1.0))
+* Rust
+  * The Ping Rate Limit type is now accessible in the Rust Language Binding ([#2528](https://github.com/mozilla/glean/pull/2528))
+  * Gracefully handle a failure when starting the upload thread. Glean no longer crashes in that case. ([#2545](https://github.com/mozilla/glean/pull/2545))
+  * `locale` now exposed through the RLB so it can be set by consumers ([2531](https://github.com/mozilla/glean/pull/2531))
+* Python
+  * Added the shutdown API for Python to ensure orderly shutdown and waiting for uploader processes ([#2538](https://github.com/mozilla/glean/pull/2538))
+* Kotlin
+  * Move running of upload task when Glean is running in a background service to use the internal Glean Dispatchers rather than WorkManager. [Bug 1844533](https://bugzilla.mozilla.org/show_bug.cgi?id=1844533)
 
 # v53.1.0 (2023-06-28)